Update __init__.py
for dhash, the file you followed http://www.hackerfactor.com/blog/index.php?/archives/529-Kind-of-Like-That.html, when you did the 3rd step that computing the difference, it said that the 9 pixels per row yields 8 differences between adjacent pixels. thus, if you print out the pixels,in my opinion, it should be a 8*9 matrix, right?